backend/pyproject.toml
[project]
name = "livekit-voice-agent"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "livekit-agents[google,openai,silero,turn-detector]~=1.3",
    "livekit-plugins-azure>=1.3.12",
    "livekit-plugins-deepgram>=1.3.12",
    "livekit-plugins-google>=1.3.12",
    "livekit-plugins-groq>=1.3.12",
    "livekit-plugins-noise-cancellation~=0.2",
    "livekit-plugins-silero>=1.3.12",
    "python-dotenv>=1.2.1",
]
